Renamed and reorganized some packages.
