+++ server/eclipse-project/src/main/java/us/freeandfair/corla/endpoint/BallotManifestDownload.java
@@ -22,8 +22,8 @@ import spark.Request;
 import spark.Response;
 
 import us.freeandfair.corla.Main;
-import us.freeandfair.corla.hibernate.Persistence;
 import us.freeandfair.corla.model.BallotManifestInfo;
+import us.freeandfair.corla.persistence.Persistence;
 import us.freeandfair.corla.util.SparkHelper;
 
 /**
