+++ server/eclipse-project/src/main/java/us/freeandfair/corla/endpoint/CVRExportUpload.java
@@ -46,14 +46,14 @@ import spark.Response;
 import us.freeandfair.corla.Main;
 import us.freeandfair.corla.csv.CVRExportParser;
 import us.freeandfair.corla.csv.DominionCVRExportParser;
-import us.freeandfair.corla.hibernate.Persistence;
 import us.freeandfair.corla.model.CastVoteRecord;
 import us.freeandfair.corla.model.CastVoteRecord.RecordType;
 import us.freeandfair.corla.model.County;
-import us.freeandfair.corla.model.CountyQueries;
 import us.freeandfair.corla.model.UploadedFile;
 import us.freeandfair.corla.model.UploadedFile.FileType;
 import us.freeandfair.corla.model.UploadedFile.HashStatus;
+import us.freeandfair.corla.persistence.Persistence;
+import us.freeandfair.corla.query.CountyQueries;
 import us.freeandfair.corla.util.FileHelper;
 import us.freeandfair.corla.util.SparkHelper;
