+++ server/eclipse-project/src/main/java/us/freeandfair/corla/model/Administrator.java
@@ -22,7 +22,7 @@ import javax.persistence.EnumType;
 import javax.persistence.Enumerated;
 import javax.persistence.Table;
 
-import us.freeandfair.corla.hibernate.AbstractEntity;
+import us.freeandfair.corla.persistence.AbstractEntity;
 
 /**
  * An administrator in the system.
