+++ server/eclipse-project/src/main/java/us/freeandfair/corla/model/County.java
@@ -25,7 +25,7 @@ import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.Table;
 
-import us.freeandfair.corla.hibernate.AbstractEntity;
+import us.freeandfair.corla.persistence.AbstractEntity;
 
 /**
  * A county involved in an audit.
