+++ server/eclipse-project/src/main/java/us/freeandfair/corla/model/UploadedFile.java
@@ -19,7 +19,7 @@ import javax.persistence.Entity;
 import javax.persistence.Lob;
 import javax.persistence.Table;
 
-import us.freeandfair.corla.hibernate.AbstractEntity;
+import us.freeandfair.corla.persistence.AbstractEntity;
 
 /**
  * An uploaded file, kept in persistent storage for archival.
