+++ server/eclipse-project/src/main/java/us/freeandfair/corla/query/DepartmentOfStateDashboardQueries.java
@@ -9,7 +9,7 @@
  * @description A system to assist in conducting statewide risk-limiting audits.
  */
 
-package us.freeandfair.corla.model;
+package us.freeandfair.corla.query;
 
 import java.util.List;
 
@@ -17,7 +17,8 @@ import javax.persistence.PersistenceException;
 import javax.persistence.RollbackException;
 
 import us.freeandfair.corla.Main;
-import us.freeandfair.corla.hibernate.Persistence;
+import us.freeandfair.corla.model.DepartmentOfStateDashboard;
+import us.freeandfair.corla.persistence.Persistence;
 
 /**
  * Queries having to do with County entities.
